requirements.txt
# Azure Functions Core
azure-functions>=1.18.0

# Salesforce Integration
simple-salesforce>=1.12.0

# Data Processing
pandas>=2.0.0

# Python Standard Library (included but listed for documentation)
# typing, json, datetime, logging are built-in
